+++ test/test_plot.py
@@ -16,7 +16,7 @@ from jinja2 import Template
 from graphtik import base, compose, network, operation, plot
 from graphtik.base import PlotArgs
 from graphtik.modifiers import optional
-from graphtik.pipeline import NetworkOperation
+from graphtik.op import NetworkOperation
 from graphtik.plot import (
     Plotter,
     Ref,